% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_notebooks.R
\name{create_notebook}
\alias{create_notebook}
\title{Create New IS Notebook}
\usage{
create_notebook(
  credentials,
  name,
  shortcut,
  initialize = TRUE,
  init_value = "",
  can_see = TRUE,
  can_add = TRUE,
  see_stat = TRUE
)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{name}{(string) a name of the new notebook}

\item{shortcut}{(string) a short-cut name of the new notebook}

\item{initialize}{(logical) whether the block should be initialized,
i.e., whether the students should be added}

\item{init_value}{(string) what string should be written to the newly
created notebook (default is "")}

\item{can_see}{(logical) whether the block can be seen by students}

\item{can_add}{(logical) whether new students can be added to the notebook}

\item{see_stat}{(logical) whether students can see statistics of the notebook}
}
\value{
none; it only creates a block in IS
}
\description{
\code{create_notebook()} creates a new IS notebook
for a course given in credentials.
}
\examples{
\dontrun{
create_notebook(micprez, "Test Block", "TB")
}

}
