% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials}
\alias{credentials}
\title{Create MUNI IS Credentials}
\usage{
credentials(key, faculty, course)
}
\arguments{
\item{key}{(string) key to IS API}

\item{faculty}{(string or numeric) faculty id number}

\item{course}{(string) course shortcut}
}
\value{
S3 object of class "IScredentials"
}
\description{
\code{credentials()} returns an object that stores all that is
necessary to access a course in IS.
}
\examples{
\dontrun{
# Microeconomics 1 at ESF MU
mic <- credentials(key = "Z6o4VCwTOPQYGWI9",
                   faculty = 1456,
                   course = "BPE_MIE1")
}
}
