% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_answers.R
\name{get_answers}
\alias{get_answers}
\title{Gets answers from a test (odpovednik) from IS.}
\usage{
get_answers(test_url, cred, type = "simple", format = "csv", ...)
}
\arguments{
\item{test_url}{A string. The URL of the test .qdesc file.
It may be shortened---you may omit "https://is.muni.cz/auth"
and start with the leading "/el/" part.}

\item{cred}{The UCO credentials.}

\item{type}{A string. The type of the output. Default is "simple".}

\item{format}{A string. The format of the output. Default is "csv".}

\item{...}{Additional parameters for the URL query.}
}
\description{
\code{get_answers()} returns a string with the content of the test answers.

\code{get_answers_table()} returns a tibble with the content of the test answers.
}
\details{
This function retrieves answers from a specified test in IS.
It requires valid UCO credentials and test URL.
}
\examples{
\dontrun{
get_answers(
 "https://is.muni.cz/auth/el/econ/podzim2023/MPE_MIVS/odp/tb/kviz/kviztema11.qdesc",
 cred
)
get_answers_table(
  "/el/econ/podzim2023/MPE_MIVS/odp/tb/kviz/kviztema11.qdesc",
  cred
)
}
}
