% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_on_seminars_and_people.R
\name{get_seminar_students}
\alias{get_seminar_students}
\title{Get students enrolled in given seminars}
\usage{
get_seminar_students(credentials, seminars = NULL)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{seminars}{An optional vector of seminar names. If given, the
function returns the name, and UČO of the students enrolled into
the given seminars; otherwise, it returns it all open seminars.}
}
\value{
A tibble of students enrolled in the given seminars.
}
\description{
\code{get_seminar_students()} returns a tibble of students
enrolled in given seminars.
}
\note{
There is some info on studium (faculty, program, ..., and
whether it is open or closed), see
https://is.muni.cz/auth/napoveda/technicka/bloky_api?fakulta=1456;obdobi=7903;predmet=1313283#predmet-info
}
\examples{
\dontrun{
get_seminar_students(cred)
get_seminar_students(cred, c("01", "02"))
}

}
