% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_on_seminars_and_people.R
\name{get_students}
\alias{get_students}
\title{Get list of all students enrolled in a course}
\usage{
get_students(credentials)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials created
by credentials()}
}
\value{
a tibble with columns uco, first_name, and last_name
}
\description{
\code{get_students()} returns a tibble of students
enrolled in a given course.
}
\note{
There is some info on studium (faculty, program, ..., and
whether it is open or closed), see
\url{https://is.muni.cz/auth/napoveda/technicka/bloky_api?#predmet-seznam}

was \code{get_list_of_all_students()}
}
\examples{
\dontrun{
get_students(cred)
}

}
