% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_notebooks.R
\name{list_notebooks}
\alias{list_notebooks}
\title{Get List of Existing IS Notebooks}
\usage{
list_notebooks(credentials)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}
}
\value{
A tibble of existing notebooks with columns "name" and "shortcut".
}
\description{
\code{list_notebooks()} returns a tibble of existing IS notebook.
}
\note{
was \code{get_list_of_is_blocks()}
}
\examples{
\dontrun{
list_notebooks(mivs)
}

}
