% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_presence.R
\name{list_presence_notebooks}
\alias{list_presence_notebooks}
\title{Get Names of All Presence Notebooks}
\usage{
list_presence_notebooks(credentials)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}
}
\value{
a tibble of names of all existing blocks; it includes columns:
name, shortcut, course, and credential
}
\description{
\code{list_presence_notebooks()} downloads
names of all existing presence blocks within several courses
}
\note{
It is protected against errors and logs them.
}
\examples{
\dontrun{
blocks <- list_presence_notebooks(micprez, mivs)
}

}
