% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_notebooks.R
\name{notebook_exists}
\alias{notebook_exists}
\title{Check that notebook exists.}
\usage{
notebook_exists(credentials, shortcut)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{shortcut}{(string) notebook shortcut name}
}
\value{
logical scalar
}
\description{
\code{notebook_exists()} checks whether a notebook exists in a course.
}
\examples{
\dontrun{
notebook_exists(mivs, "bodysemin02")
}

}
