% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_presence.R
\name{read_all_presence_points}
\alias{read_all_presence_points}
\title{Read presence points of all students}
\usage{
read_all_presence_points(credentials)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}
}
\value{
a tibble which includes two columns: uco and number of points
}
\description{
\code{read_all_presence_points()} returns a tibble
containing the number of attendances of all students enrolled
in several courses.
}
\note{
It is protected against errors; it logs.
}
\examples{
\dontrun{
pts <- read_all_presence_points(micprez, mivs)
}

}
