% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_notebooks.R
\name{read_notebook}
\alias{read_notebook}
\title{Read content of an IS notebook}
\usage{
read_notebook(credentials, notebook)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{notebook}{(string) the shortcut of the notebook}
}
\value{
A tibble with columns "uco", "content", "modified",
and "credentials".
}
\description{
\code{read_notebook()} reads the content of an IS notebook and returns it
as a tibble.
}
\note{
was \code{read_is_block()}

Problem: All students have the same modified date. I don't know why.
}
\examples{
\dontrun{
read_notebook(mivs, "test")
}

}
