% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_notebooks.R
\name{read_point_notebook}
\alias{read_point_notebook}
\title{Read Points from IS Notebook}
\usage{
read_point_notebook(credentials, notebook)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{notebook}{(string) the short name of the notebook}
}
\value{
a tibble with columns uco and points
}
\description{
\code{read_point_notebook()} reads IS notebook and
returns points stored there as a tibble.
}
\note{
The function extracts only UČO and points;
the points are expected to be in the following formats:
"*11", "*1.2", "*1,3", "*0.5", etc.

was \code{read_is_point_block()}
}
\examples{
\dontrun{
read_point_notebook(mivs, "test2")
}

}
