% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_presence.R
\name{read_presence_notebook}
\alias{read_presence_notebook}
\title{Read Students' Presence from IS}
\usage{
read_presence_notebook(credentials, notebook)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{notebook}{(string) shortcut of the presence notebook}
}
\value{
A tibble with columns "uco", "content", "date", and "attendance".
}
\description{
\code{read_presence_notebook()} reads IS presence notebook and returns
its values as a tibble.
}
\note{
The function extracts UČO, content, number of attendances and
number of missed seminars; it is supposed that the content is in form
"DD.(M)M.: (ne)účast".
}
\examples{
\dontrun{
read_presence_notebook(mivs, "test2")
}

}
