% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uco_credentials.R
\name{uco_credentials}
\alias{uco_credentials}
\title{Create MUNI IS UCO Credentials}
\usage{
uco_credentials(uco, password)
}
\arguments{
\item{uco}{(string) non-personal UCO}

\item{password}{(string) password for the UCO}
}
\value{
S3 object of class "ISUCOCredentials"
}
\description{
\code{uco_credentials()} returns an object that stores all that is
necessary to access files in IS.
}
\note{
See https://is.muni.cz/auth/elearning/test_export?help=1.
}
\examples{
\dontrun{
# Microeconomics 1 at ESF MU
cred <- uco_credentials(
   uco = "123456",
  password = "password"
)
}
}
