% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_notebooks.R
\name{write_notebook}
\alias{write_notebook}
\title{Write table of values to IS notebook}
\usage{
write_notebook(credentials, notebook, tab, overwrite = TRUE)
}
\arguments{
\item{credentials}{(object of class "IScredentials") credentials
created by credentials()}

\item{notebook}{(string) a short name of an IS notebook}

\item{tab}{(tibble) what should be written and for which students; it must
contain the following columns:
\itemize{
\item uco (string) students' ucos
\item value (string) the content that should be written there
}}

\item{overwrite}{(logical scalar) whether the content should be overwritten
(default is TRUE)}
}
\value{
none; it only writes to IS
}
\description{
\code{write_notebook()} writes content to the IS block.
}
\examples{
\dontrun{
tab <- students |>
    filter(course == "mivs") |>
    select(uco = student_uco, value = full_string)
write_notebook(mivs, "test2", tab)
}

}
